requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
attrs==24.3.0
autogen==0.3.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
deprecated==1.2.15
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
email-validator==2.2.0
eval-type-backport==0.2.2
executing==2.1.0
fastapi==0.111.1
fastapi-cli==0.0.5
filelock==3.16.1
flaml==2.3.1
frozenlist==1.5.0
fsspec==2024.9.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
greenlet==3.0.3
griffe==1.5.4
groq==0.13.1
grpcio==1.67.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.10
importlib-metadata==8.5.0
jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpath-python==1.0.6
logfire==2.11.0
logfire-api==2.11.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mistralai==1.2.6
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==1.59.3
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pdfminer-six==20231228
pdfplumber==0.11.1
pillow==11.0.0
playwright==1.44.0
propcache==0.2.1
proto-plus==1.24.0
protobuf==5.29.2
psutil==6.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyautogen==0.2.27
pycparser==2.22
pydantic==2.10.4
pydantic-ai==0.0.17
pydantic-ai-slim==0.0.17
pydantic-core==2.27.2
pyee==11.1.0
pygments==2.18.0
pyparsing==3.2.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.11
pytz==2024.2
pyyaml==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
tabulate==0.9.0
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.67.1
typer==0.12.5
typing-extensions==4.12.2
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.3
watchfiles==0.24.0
websockets==13.1
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
